add option to save :tells after they are sent

+++ classes/xmppEntity.php
@@ -178,7 +178,7 @@ class xmppEntity extends ligrevGlobals {
     if ($_ligrevStartupInhibitTell) {
       return false;
     }
-    $sql   = $this->db->prepare('SELECT * FROM tell WHERE recipient = ? ORDER BY sent ASC',
+    $sql   = $this->db->prepare('SELECT * FROM tell WHERE recipient = ? and isDelivered = 0 ORDER BY sent ASC',
       array("string"));
     $sql->bindValue(1, str_replace("\\20", " ", $this->jid->bare), "string");
     $sql->execute();
@@ -199,7 +199,11 @@ class xmppEntity extends ligrevGlobals {
       } else {
         $this->sendMessage($room, $message, true, "groupchat");
       }
-      $this->db->delete('tell', ['id' => $tell['id']]);
+      if ($this->config['archiveTells']) {
+        $this->db->update('tell', ['isDelivered' => 1], ['id' => $tell['id']]);
+      } else {
+        $this->db->delete('tell', ['id' => $tell['id']]);
+      }
     }
   }
